+++ tools/lgtdoc/xml/lgt2html.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   XML documenting files to (X)HTML conversion script 
-##   Last updated on April 13, 2022
+##   Last updated on July 18, 2022
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2022 Paulo Moura <pmoura@logtalk.org>
@@ -100,7 +100,6 @@ index_title="Documentation index"
 
 processor=xsltproc
 # processor=xalan
-# processor=sabcmd
 # processor=saxon
 
 usage_help()
@@ -119,7 +118,7 @@ usage_help()
 	echo "  -d output directory for the generated files (default is $directory)"
 	echo "  -i name of the index file (default is $index_file)"
 	echo "  -t title to be used in the index file (default is $index_title)"
-	echo "  -p XSLT processor (xsltproc, xalan, sabcmd, or saxon; default is $processor)"
+	echo "  -p XSLT processor (xsltproc, xalan, or saxon; default is $processor)"
 	echo "  -v print version"
 	echo "  -h help"
 	echo
@@ -219,7 +218,7 @@ if [ "$t_arg" != "" ] ; then
 	index_title=$t_arg
 fi
 
-if [ "$p_arg" != "" ] && [ "$p_arg" != "xsltproc" ] && [ "$p_arg" != "xalan" ] && [ "$p_arg" != "sabcmd" ] && [ "$p_arg" != "saxon" ] ; then
+if [ "$p_arg" != "" ] && [ "$p_arg" != "xsltproc" ] && [ "$p_arg" != "xalan" ] && [ "$p_arg" != "saxon" ] ; then
 	echo "Error! Unsupported XSLT processor: $p_arg" >&2
 	usage_help
 	exit 1
@@ -269,7 +268,6 @@ if grep -q "<logtalk" ./*.xml ; then
 		case "$processor" in
 			xsltproc)	eval xsltproc -o \"$directory/$name.html\" \"$entity_xslt\" \"$file\";;
 			xalan)		eval xalan -o \"$directory/$name.html\" \"$file\" \"$entity_xslt\";;
-			sabcmd)		eval sabcmd \"$entity_xslt\" \"$file\" \"$directory/$name.html\";;
 			saxon)		eval java net.sf.saxon.Transform -o:\"$directory/$name.html\" -s:\"$file\" -xsl:\"$entity_xslt\";;
 		esac
 	done
@@ -280,7 +278,6 @@ if grep -q "<logtalk" ./*.xml ; then
 		case "$processor" in
 			xsltproc)	eval xsltproc -o \"$directory/$name.html\" \"$index_xslt\" \"$file\";;
 			xalan)		eval xalan -o \"$directory/$name.html\" \"$file\" \"$index_xslt\";;
-			sabcmd)		eval sabcmd \"$index_xslt\" \"$file\" \"$directory/$name.html\";;
 			saxon)		eval java net.sf.saxon.Transform -o:\"$directory/$name.html\" -s:\"$file\" -xsl:\"$index_xslt\";;
 		esac
 	done